Add files via upload
Business Ticker token is a store of value built on the Ethereum blockchain with a unique deflationary features; which will be used to bring about it users records around the globe, to store series of data's and to keep intensive history of all blockchain technology in the past and also in making.
BLOCKCHAIN OVERVIEW
All components related to Business Tickerblockchain platform will run on the Ethereum network. There are two co-dependent facilitators of all interactions on the blockchain detailed below:

Smart Contracts and the Business Tickertoken. Ethereum is different from Bitcoin because it allows you to create smart contracts that can be described as highly programmable digital money. Contrary to many other cryptocurrencies, Ethereum is not only a network for the exchange of monetary value but a network to run contracts based on Ethereum. These contracts can be used safely to perform a large number of operations: electoral systems, registration of domain names, financial markets, crowdfunding platforms, intellectual property, etc. Business Tickerreserves the right to use one or more blockchains within its project. This will allow us to be totally decentralized without depending on a single technology.